ci fix: Do not fixate DISKUV_OPAM_REPOSITORY
It is better to let the latest version of dkml-workflows use its DEFAULT_DISKUV_OPAM_REPOSITORY_TAG which will be compatible (ie. almost always the same) with DEFAULT_DKML_COMPILER and more important the pins will all exist in the (default tag) diskuv-opam-repository.

It is too easy, especially because diskuv-opam-repository is now pruned to remove unused package versions, for some pins not to exist when DISKUV_OPAM_REPOSITORY was incorrect.